% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel2.R
\name{read_excel2}
\alias{read_excel2}
\title{Read xls and xlsx files.}
\usage{
read_excel2(path, sheet = 1, col_names = TRUE, col_types = NULL,
  na = "", skip = 0, n_max = 1050000L)
}
\arguments{
\item{path}{Path to the xls/xlsx file}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or
an integer (the position of the sheet). Defaults to the first sheet.}

\item{col_names}{Either \code{TRUE} to use the first row as column names,
\code{FALSE} to number columns sequentially from \code{X1} to \code{Xn},
or a character vector giving a name for each column.}

\item{col_types}{Either \code{NULL} to guess from the spreadsheet or a
character vector containing "blank", "numeric", "date" or "text".}

\item{na}{Missing value. By default readxl converts blank cells to missing
data. Set this value if you have used a sentinel value for missing values.}

\item{skip}{Number of rows to skip before reading any data.}

\item{n_max}{Max number of rows to evalutate for internal function xls_col_types and xlsx_col_types. Defaults to all rows (1050000L)}
}
\description{
Import excel files into R similar to readxl::read_excel,
  but with all rows evaluated in xls_col_types and xlsx_col_types
}
\seealso{
\code{\link[readxl]{read_excel}}
}

